feat(RHINENG-10417): rename groups to workspaces
