+++ package.json
@@ -62,6 +62,7 @@
     "@redhat-cloud-services/frontend-components-notifications": "^4.1.0",
     "@redhat-cloud-services/frontend-components-translations": "^3.2.7",
     "@redhat-cloud-services/frontend-components-utilities": "^4.0.2",
+    "@unleash/proxy-client-react": "^4.2.4",
     "classnames": "2.3.1",
     "cypress": "^12.0.2",
     "graphql": "15.5.3",
