+++ src/Components/SysTable/SysTable.js
@@ -7,7 +7,7 @@ import {
     useFetchSystems,
     buildApiFilters
 } from './hooks';
-import { defaultOnLoad, columns, mergedColumns, appendDirection, findColumnByKey } from './constants';
+import { defaultOnLoad, columns, appendDirection, findColumnByKey } from './constants';
 
 import { InventoryTable } from '@redhat-cloud-services/frontend-components/Inventory';
 import { GET_SYSTEM_TABLE } from '../../operations/queries';
@@ -22,6 +22,7 @@ import { globalFilters } from '../../store/cache';
 import { useDispatch } from 'react-redux';
 import  matchFilter  from './matchFilter';
 import { buildActiveFiltersConfig } from './Helpers';
+import { useWorkspaceFeatureFlag } from '../../Utilities/Hooks';
 
 const SysTable = () => {
     const inventory = useRef(null);
@@ -38,6 +39,29 @@ const SysTable = () => {
     const fetchSystems = useCallback(useFetchSystems({
         query: GET_SYSTEM_TABLE
     }), []);
+    const isWorkspaceEnabled = useWorkspaceFeatureFlag()
+
+    const mergedColumns = (defaultColumns) => {
+        return columns(isWorkspaceEnabled).map((column) => {
+            const isStringCol = typeof column === 'string';
+
+            const defaultColumn = defaultColumns.find(
+                (defaultCol) => defaultCol.key === column.key
+            );
+            //this step is necessary so we could compare the default columns from column object at the top of the file
+            //with the columns that come from the inventory, so if inventory hide a column behind feature flag
+            //it will not appear when the flag is turned off
+            return column.requiresDefault && defaultColumn === undefined ?
+                undefined : {
+                    ...(isStringCol ? { key: column } : column),
+                    ...defaultColumn,
+                    props: {
+                        ...defaultColumn?.props,
+                        ...column?.props
+                    }
+                };
+        }).filter(Boolean);
+    };
 
     const { value } = matchFilter(matchStateValue, setStateValue)?.filterValues;
     return <React.Fragment>
@@ -56,7 +80,7 @@ const SysTable = () => {
             hasCheckbox={false}
             showTags
             ref={inventory}
-            onLoad={defaultOnLoad(columns, getRegistry)}
+            onLoad={defaultOnLoad(columns(isWorkspaceEnabled), getRegistry)}
             getEntities={async (
                 _ids,
                 { page = 1, per_page: perPage, orderBy, orderDirection, filters }
