+++ src/Components/SysTable/constants.js
@@ -10,6 +10,7 @@ import StatusLabel from '../StatusLabel/StatusLabel';
 import { Tooltip } from '@patternfly/react-core/dist/esm/components/Tooltip/Tooltip';
 import { DateFormat } from '@redhat-cloud-services/frontend-components/DateFormat';
 import { InsightsLink } from '@redhat-cloud-services/frontend-components/InsightsLink';
+import { useWorkspaceFeatureFlag } from '../../Utilities/Hooks';
 
 const cache = createIntlCache();
 const locale = navigator.language.slice(0, 2);
@@ -34,7 +35,7 @@ export const defaultOnLoad =
                   })
           });
 
-export const columns = [
+export const columns = (workspacesEnabled) => [
     {
         title: intl.formatMessage(messages.name),
         transforms: [cellWidth(30)],
@@ -43,7 +44,7 @@ export const columns = [
         renderFunc: (displayName, id) => <InsightsLink to={`/systems/${id}`}>{displayName}</InsightsLink>
     },
     {
-        title: 'Group',
+        title: workspacesEnabled ? 'Workspace' : 'Group',
         key: 'groups',
         sortBy: ['GROUP_NAME'],
         requiresDefault: true,
@@ -105,28 +106,6 @@ export const columns = [
 
     }
 ];
-
-export const mergedColumns = (defaultColumns) => {
-    return columns.map((column) => {
-        const isStringCol = typeof column === 'string';
-
-        const defaultColumn = defaultColumns.find(
-            (defaultCol) => defaultCol.key === column.key
-        );
-        //this step is necessary so we could compare the default columns from column object at the top of the file
-        //with the columns that come from the inventory, so if inventory hide a column behind feature flag
-        //it will not appear when the flag is turned off
-        return column.requiresDefault && defaultColumn === undefined ?
-            undefined : {
-                ...(isStringCol ? { key: column } : column),
-                ...defaultColumn,
-                props: {
-                    ...defaultColumn?.props,
-                    ...column?.props
-                }
-            };
-    }).filter(Boolean);
-};
 
 export const appendDirection = (attributes, direction) =>
     attributes?.map((attribute) => `${attribute}_${direction}`);
